+++ src/main/java/com/telepathicgrunt/blame/Blame.java
@@ -1,9 +1,5 @@
 package com.telepathicgrunt.blame;
 
-import net.minecraft.world.gen.GenerationStage;
-import net.minecraft.world.gen.feature.Feature;
-import net.minecraft.world.gen.feature.NoFeatureConfig;
-import net.minecraftforge.event.world.BiomeLoadingEvent;
 import net.minecraftforge.fml.ModList;
 import net.minecraftforge.fml.common.Mod;
 import org.apache.logging.log4j.Level;
@@ -15,7 +11,7 @@ public class Blame
 {
     public static final String MODID = "blame";
     public static final Logger LOGGER = LogManager.getLogger();
-    public static String VERSION = "1.7.4";
+    public static String VERSION = "1.7.5";
 
     public Blame() {
         ModList.get().getModContainerById(Blame.MODID)
@@ -23,13 +19,16 @@ public class Blame
 
         Blame.LOGGER.log(Level.ERROR, "Blame "+VERSION+" initialized");
 
+        // Test detecting dispenser registry replacements
+        // DispenserBlock.registerDispenseBehavior(Items.HONEY_BOTTLE, new DefaultDispenseItemBehavior());
+
         // Test detecting unregistered configuredfeatures.
         // MinecraftForge.EVENT_BUS.addListener(EventPriority.HIGH, this::biomeModification);
     }
     
-    public void biomeModification(final BiomeLoadingEvent event) {
-        // Add our structure to all biomes including other modded biomes.
-        // You can filter to certain biomes based on stuff like temperature, scale, precipitation, mod id.
-        event.getGeneration().getFeatures(GenerationStage.Decoration.UNDERGROUND_ORES).add(() -> Feature.DESERT_WELL.withConfiguration(new NoFeatureConfig()));
-    }
+//    public void biomeModification(final BiomeLoadingEvent event) {
+//        // Add our structure to all biomes including other modded biomes.
+//        // You can filter to certain biomes based on stuff like temperature, scale, precipitation, mod id.
+//        event.getGeneration().getFeatures(GenerationStage.Decoration.UNDERGROUND_ORES).add(() -> Feature.DESERT_WELL.withConfiguration(new NoFeatureConfig()));
+//    }
 }