+++ src/main/java/com/telepathicgrunt/blame/main/DispenserBlockRegistry.java
@@ -3,6 +3,7 @@ package com.telepathicgrunt.blame.main;
 import com.telepathicgrunt.blame.Blame;
 import it.unimi.dsi.fastutil.objects.Object2ObjectOpenHashMap;
 import net.minecraft.item.Item;
+import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.registry.Registry;
 import org.apache.logging.log4j.Level;
 
@@ -16,18 +17,22 @@ import static it.unimi.dsi.fastutil.HashCommon.arraySize;
  */
 public class DispenserBlockRegistry<K, V> extends Object2ObjectOpenHashMap<K, V>{
 
+	// Turn on registry replacement detection only after startup's putAll I do is done.
+	public Boolean startupIgnore = true;
+
 	@Override
 	public V put(final K item, final V behavior) {
-
-		if(this.containsKey(item)){
+		// Can't use Forge Item Registry as it is null when this method is first called by vanilla.
+		ResourceLocation itemRl = Registry.ITEM.getKey((Item)item);
+		if(!startupIgnore && (itemRl.getNamespace().equals("minecraft") || this.containsKey(item))){
 			StackTraceElement stack = Thread.currentThread().getStackTrace()[3];
 			StackTraceElement stack2 = Thread.currentThread().getStackTrace()[4];
 			StackTraceElement stack3 = Thread.currentThread().getStackTrace()[5];
-			Blame.LOGGER.log(Level.ERROR,"\n****************** Blame Extra Info Report " + Blame.VERSION + " ******************" +
+			Blame.LOGGER.log(Level.ERROR, "\n****************** Blame Extra Info Report " + Blame.VERSION + " ******************" +
 					"\n   Ignore this unless item behavior aren't working with Dispensers. If Dispenser behavior" +
 					"\n   is broken, check out \"Potentially Dangerous alternative prefix `minecraft`\" lines for" +
 					"\n   the item too as registry replacements might break dispenser behaviors as well." +
-					"\n  Dispenser Behavior overridden for " + Registry.ITEM.getKey((Item)item).toString() +
+					"\n  Dispenser Behavior overridden for " + itemRl.toString() +
 					"\n  New behavior: " + behavior.getClass().getName() +
 					"\n  Old behavior: " + this.get(item).getClass().getName() +
 					"\n  Registration done at: " +